+++ src/renderer/public/themes/light.json
@@ -1,0 +1,26 @@
+{
+  "primary": "#0C4BEE",
+  "primaryHover": "#225EF9",
+  "primaryActive": "#2A62F4",
+  "error": "red",
+  "success": "lime",
+
+  "bg-primary": "#e7e7e7",
+  "bg-secondary": "#f2f2f2",
+  "bg-hover": "#cfcfcf",
+  "bg-active": "#c2c2c2",
+
+  "border-primary": "#c2c2c2",
+
+  "text-primary": "#000",
+  "text-secondary": "#fff",
+  "text-inactive": "#b3b2b2",
+  "text-disabled": "rgb(156,163,175)",
+
+  "codeEditorTheme": "vs-light",
+
+  "scrollbar-track": "rgba(162,162,162, 0.5)",
+  "scrollbar-thumb": "rgba(162,162,162, 0.7)",
+
+  "grid": "rgba(0,0,0,0.08)"
+}
