+++ src/renderer/src/components/Compiler.tsx
@@ -1,32 +1,107 @@
-import { CompilerResult } from '@renderer/types/CompilerTypes';
-import React from 'react';
+import React, { Dispatch, SetStateAction, useEffect, useState } from 'react';
 import { twMerge } from 'tailwind-merge';
 
+import { useSidebar } from '@renderer/store/useSidebar';
+import { CompilerResult } from '@renderer/types/CompilerTypes';
+
+import { Compiler } from '../components/Modules/Compiler';
+import { Settings } from '../components/Modules/Settings';
+import { useTabs } from '@renderer/store/useTabs';
+import { EditorData, EditorManager } from '@renderer/lib/data/EditorManager';
+import { CanvasEditor } from '@renderer/lib/CanvasEditor';
+
 export interface CompilerProps {
+  manager: EditorManager | null;
+  editor: CanvasEditor | null;
+  editorData: EditorData;
+
+  openData: [boolean, string | null, string | null, string] | undefined;
   compilerData: CompilerResult | undefined;
+  setCompilerData: Dispatch<SetStateAction<CompilerResult | undefined>>;
   compilerStatus: string;
-  fileReady: boolean;
-  handleCompile: () => void;
-  handleAddStdoutTab: () => void;
-  handleAddStderrTab: () => void;
-  handleSaveSourceIntoFolder: () => void;
-  handleSaveBinaryIntoFolder: () => void;
-  handleShowSource: () => void;
-  handleFlashButton: () => void;
+  setCompilerStatus: Dispatch<SetStateAction<string>>;
 }
 
-export const Compiler: React.FC<CompilerProps> = ({
+export const CompilerTab: React.FC<CompilerProps> = ({
+  manager,
+  editor,
+  editorData,
+  openData,
   compilerData,
+  setCompilerData,
   compilerStatus,
-  fileReady,
-  handleCompile,
-  handleAddStdoutTab,
-  handleAddStderrTab,
-  handleSaveSourceIntoFolder,
-  handleSaveBinaryIntoFolder,
-  handleShowSource,
-  handleFlashButton,
+  setCompilerStatus,
 }) => {
+  const openTab = useTabs((state) => state.openTab);
+  const changeSidebarTab = useSidebar((state) => state.changeTab);
+
+  const [importData, setImportData] = useState<string | undefined>(undefined);
+
+  const handleFlashButton = () => {
+    // TODO: индекс должен браться из какой-то переменной
+    changeSidebarTab(3);
+  };
+
+  const handleSaveBinaryIntoFolder = async () => {
+    const preparedData = await Compiler.prepareToSave(compilerData!.binary!);
+    manager?.saveIntoFolder(preparedData);
+  };
+
+  const handleCompile = async () => {
+    Compiler.filename = editorData.shownName!;
+    manager?.compile(editor!.container.machine.platformIdx);
+  };
+
+  const handleSaveSourceIntoFolder = async () => {
+    await manager?.saveIntoFolder(compilerData!.source!);
+  };
+
+  const handleAddStdoutTab = () => {
+    openTab({
+      type: 'code',
+      name: 'stdout',
+      code: compilerData!.stdout ?? '',
+      language: 'txt',
+    });
+  };
+
+  const handleAddStderrTab = () => {
+    openTab({
+      type: 'code',
+      name: 'stderr',
+      code: compilerData!.stderr ?? '',
+      language: 'txt',
+    });
+  };
+
+  const handleShowSource = () => {
+    compilerData!.source!.forEach((element) => {
+      openTab({
+        type: 'code',
+        name: `${element.filename}.${element.extension}`,
+        code: element.fileContent,
+        language: 'cpp',
+      });
+    });
+  };
+
+  useEffect(() => {
+    if (importData && openData) {
+      manager?.parseImportData(importData, openData!);
+      setImportData(undefined);
+    }
+  }, [importData]);
+
+  useEffect(() => {
+    Compiler.bindReact(setCompilerData, setCompilerStatus, setImportData);
+    Settings.getCompilerSettings().then((compiler) => {
+      console.log('CONNECTING TO COMPILER');
+      Compiler.connect(compiler.host, compiler.port);
+    });
+  }, []);
+
+  const fileReady = !!editor;
+
   const button = [
     {
       name: 'Показать stderr',
