+++ src/renderer/src/components/Sidebar/Sidebar.tsx
@@ -1,21 +1,11 @@
-import React, { useMemo } from 'react';
+import React, { useMemo, useState } from 'react';
 
-import {
-  Explorer,
-  Menu,
-  Compiler,
-  Loader,
-  MenuProps,
-  CompilerProps,
-  FlasherProps,
-  ExplorerCallbacks,
-} from '../../components';
+import { Explorer, Menu, CompilerTab, Loader, ExplorerCallbacks } from '../../components';
 
 import { ReactComponent as MenuIcon } from '@renderer/assets/icons/menu.svg';
 import { ReactComponent as ComponentsIcon } from '@renderer/assets/icons/components.svg';
 import { ReactComponent as CompilerIcon } from '@renderer/assets/icons/compiler.svg';
 import { ReactComponent as DriveIcon } from '@renderer/assets/icons/drive.svg';
-//import { ReactComponent as ChipIcon } from '@renderer/assets/icons/chip.svg';
 import { ReactComponent as SettingsIcon } from '@renderer/assets/icons/settings.svg';
 import { Setting } from '../Setting';
 import { EditorRef } from '@renderer/hooks/useEditorManager';
@@ -23,6 +13,7 @@ import { EditorRef } from '@renderer/hooks/useEditorManager';
 import { Labels } from './Labels';
 import { Menus } from './Menus';
 import { Badge } from '../UI';
+import { CompilerResult } from '@renderer/types/CompilerTypes';
 
 export interface SidebarCallbacks {
   onRequestNewFile: () => void;
@@ -31,20 +22,15 @@ export interface SidebarCallbacks {
   onRequestSaveAsFile: () => void;
   onRequestAddComponent: () => void;
   onRequestEditComponent: (idx: string) => void;
-  onRequestImport: (platform: string) => void;
   onRequestDeleteComponent: (name: string) => void;
 }
 
 interface SidebarProps {
   editorRef: EditorRef;
   callbacks: SidebarCallbacks;
-  compilerProps: CompilerProps;
-  flasherProps: FlasherProps;
 }
 
 export const Sidebar: React.FC<SidebarProps> = ({
-  flasherProps,
-  compilerProps,
   editorRef,
   callbacks: {
     onRequestNewFile,
@@ -54,9 +40,52 @@ export const Sidebar: React.FC<SidebarProps> = ({
     onRequestAddComponent,
     onRequestEditComponent,
     onRequestDeleteComponent,
-    onRequestImport,
   },
 }) => {
+  const [openData, setOpenData] = useState<
+    [boolean, string | null, string | null, string] | undefined
+  >(undefined);
+  const [compilerData, setCompilerData] = useState<CompilerResult | undefined>(undefined);
+  const [compilerStatus, setCompilerStatus] = useState('Не подключен.');
+
+  const handleImport = async (platform: string) => {
+    await editorRef.managerRef.current?.import(platform, setOpenData);
+  };
+
+  const menus = useMemo(
+    () => [
+      <Menu
+        onRequestNewFile={onRequestNewFile}
+        onRequestOpenFile={onRequestOpenFile}
+        onRequestSaveFile={onRequestSaveFile}
+        onRequestSaveAsFile={onRequestSaveAsFile}
+        onRequestImport={handleImport}
+        compilerStatus={compilerStatus}
+      />,
+      <Explorer
+        editorRef={editorRef}
+        callbacks={{
+          onRequestAddComponent,
+          onRequestEditComponent,
+          onRequestDeleteComponent,
+        }}
+      />,
+      <CompilerTab
+        manager={editorRef.managerRef.current}
+        editor={editorRef.editor}
+        editorData={editorRef.editorData}
+        openData={openData}
+        compilerData={compilerData}
+        setCompilerData={setCompilerData}
+        compilerStatus={compilerStatus}
+        setCompilerStatus={setCompilerStatus}
+      />,
+      <Loader manager={editorRef.managerRef.current} compilerData={compilerData} />,
+      <Setting />,
+    ],
+    [editorRef]
+  );
+
   const tabLabels = useMemo(
     () => [
       {
@@ -75,41 +104,12 @@ export const Sidebar: React.FC<SidebarProps> = ({
       {
         Icon: <DriveIcon />,
       },
-      // {
-      //   Icon: <ChipIcon />,
-      // },
       {
         Icon: <SettingsIcon />,
         bottom: true,
       },
     ],
     [editorRef.editorData]
-  );
-
-  const menuProps: MenuProps = {
-    onRequestNewFile,
-    onRequestOpenFile,
-    onRequestSaveFile,
-    onRequestSaveAsFile,
-    onRequestImport,
-    compilerStatus: compilerProps.compilerStatus,
-  };
-
-  const explorerCallbacks: ExplorerCallbacks = {
-    onRequestAddComponent,
-    onRequestEditComponent,
-    onRequestDeleteComponent,
-  };
-
-  const menus = useMemo(
-    () => [
-      <Menu {...menuProps} />,
-      <Explorer editorRef={editorRef} callbacks={explorerCallbacks} />,
-      <Compiler {...compilerProps} />,
-      <Loader {...flasherProps} />,
-      <Setting />,
-    ],
-    [editorRef.editorData, compilerProps, flasherProps]
   );
 
   return (
